use int for markets, use datetime in setup and run - removes need for closeness checks due to floating point

+++ assume/units/demand.py
@@ -25,7 +25,7 @@ class Demand(BaseUnit):
         id: str,
         technology: str = None,
         node: str = None,
-        price: float = 900,
+        price: int = 900,
         volume: int = -1000,
         location: tuple[float, float] = None,
         bidding_strategy: BaseStrategy = {},
